requirements.txt
# Python dependencies for Personal AI Employee

# Core dependencies
python-dotenv==1.0.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Browser automation
playwright==1.41.0

# File system monitoring
watchdog==3.0.0

# Process management
psutil==5.9.8

# HTTP requests
requests==2.31.0

# Logging and utilities
python-dateutil==2.8.2
